[ISSUE-32] move: 페이지 루트 별 컴포넌트 폴더 생성 및 경로 수정 (#26)
