+++ src/app/(main)/course/components/MapView.jsx
