+++ src/app/(main)/course/components/ShelterCardComponent.jsx
@@ -3,13 +3,13 @@
 import Image from 'next/image';
 import styled from 'styled-components';
 
-import bookmarkIcon from '../../../assets/icons/bookmark-icon.png';
-import callIcon from '../../../assets/icons/call-icon.png';
-import shareIcon from '../../../assets/icons/share-post.png';
-import shelterImg from '../../../assets/images/tempShelter.png';
-import { TEXT_COLORS } from '../../../constants/colors';
-import { FONTS } from '../../../constants/font';
-import { PADDING_HORIZONTAL } from '../../../constants/space';
+import bookmarkIcon from '../../../../assets/icons/bookmark-icon.png';
+import callIcon from '../../../../assets/icons/call-icon.png';
+import shareIcon from '../../../../assets/icons/share-post.png';
+import shelterImg from '../../../../assets/images/tempShelter.png';
+import { TEXT_COLORS } from '../../../../constants/colors';
+import { FONTS } from '../../../../constants/font';
+import { PADDING_HORIZONTAL } from '../../../../constants/space';
 
 const ShelterCardComponent = () => {
     return (
