+++ src/app/(main)/home/info/page.jsx
@@ -1,4 +1,4 @@
-'use client'
+'use client';
 
 import styled from 'styled-components';
 
@@ -8,7 +8,7 @@ import CustomHeader from '../../../../components/layout/CustomHeader';
 import { BUTTON_COLORS } from '../../../../constants/colors';
 /*import { FONTS } from '../../../../constants/font';*/
 import { PADDING_HORIZONTAL, PADDING_VERTICAL } from '../../../../constants/space';
-import DogProfileComponent from '../DogProfile';
+import DogProfileComponent from '../components/DogProfile';
 
 const InfoMatch = () => {
     return (
@@ -16,7 +16,7 @@ const InfoMatch = () => {
             <CustomHeader title="정보 기반 매칭" showBackBtn={true} />
             <InfoBanner message="동행자님의 입력 정보에 맞는 멍즈를 찾았어요! 🐶" />
             <TagBox>태그가 들어갈 자리</TagBox>
-            <DogProfileComponent subtitle="나와 맞는 멍즈를 찾았어요! 🐾"/>
+            <DogProfileComponent subtitle="나와 맞는 멍즈를 찾았어요! 🐾" />
             <ContentContainer>
                 <ButtonWrapper>
                     <CustomButton color={BUTTON_COLORS.primary} text="동행 신청하기" route="/home/info" />
@@ -29,9 +29,7 @@ const InfoMatch = () => {
 
 export default InfoMatch;
 
-const TagBox = styled.div`
-    
-`
+const TagBox = styled.div``;
 const ContentContainer = styled.div`
     ${PADDING_HORIZONTAL}
     ${PADDING_VERTICAL}
