+++ src/app/(main)/home/mbti/page.jsx
@@ -1,4 +1,4 @@
-'use client'
+'use client';
 
 import styled from 'styled-components';
 
@@ -9,8 +9,8 @@ import CustomHeader from '../../../../components/layout/CustomHeader';
 import { BUTTON_COLORS } from '../../../../constants/colors';
 /*import { FONTS } from '../../../../constants/font';*/
 import { PADDING_HORIZONTAL, PADDING_VERTICAL } from '../../../../constants/space';
-import DogProfileComponent from '../DogProfile';
-import MbtiMatchComponent from './MbtiMatch';
+import DogProfileComponent from '../components/DogProfile';
+import MbtiMatchComponent from '../components/MbtiMatch';
 
 const MbtiMatch = () => {
     return (
@@ -20,8 +20,8 @@ const MbtiMatch = () => {
             <MbtiMatchContainer>
                 <MbtiMatchComponent />
             </MbtiMatchContainer>
-            <SectionDivider/>
-            <DogProfileComponent subtitle = "나와 맞는 멍즈를 찾았어요! 🐾"/>
+            <SectionDivider />
+            <DogProfileComponent subtitle="나와 맞는 멍즈를 찾았어요! 🐾" />
             <ContentContainer>
                 <ButtonWrapper>
                     <CustomButton color={BUTTON_COLORS.primary} text="동행 신청하기" route="/home/info" />
