+++ src/app/(main)/home/page.jsx
@@ -9,10 +9,9 @@ import CustomHeader from '../../../components/layout/CustomHeader';
 import { BUTTON_COLORS, TEXT_COLORS } from '../../../constants/colors';
 import { FONTS } from '../../../constants/font';
 import { PADDING_HORIZONTAL, PADDING_VERTICAL } from '../../../constants/space';
-import DogProfileComponent from './DogProfile';
+import DogProfileComponent from './components/DogProfile';
 
 const Home = () => {
-
     return (
         <Container>
             <CustomHeader />
@@ -25,7 +24,7 @@ const Home = () => {
                 </ButtonWrapper>
             </ContentContainer>
             <SectionDivider />
-            <DogProfileComponent subtitle = "오늘의 추천 멍즈 🐾" />
+            <DogProfileComponent subtitle="오늘의 추천 멍즈 🐾" />
         </Container>
     );
 };
@@ -56,4 +55,3 @@ const ButtonWrapper = styled.div`
     flex-direction: column;
     gap: 1rem;
 `;
-
